+++ src/set.js
@@ -2,18 +2,19 @@ import curry from './curry'
 import set from './uncurried/set'
 
 /**
- * Sets a property of an object.
+ * Sets a property of an object using a key path.
  *
  * @function
- * @param {String} k The key.
+ * @param {Array|String} ks The key path.
  * @param v The value to set.
  * @param {Object} o The object.
- * @returns {Object} A copy of the object `o` with the property `k` set to the
- * value `v`.
+ * @returns {Object} A copy of the object `o` with the property at the key
+ * path `ks` set to the value `v`.
  * @example
  *
  * import { set } from 'fkit'
- * const person = { name: 'Jane', age: 20, city: 'Melbourne' }
- * set('name', 'Steve', person) // { name: 'Steve', age: 20, city: 'Melbourne' }
+ * const person = { name: 'Jane', age: 20, address: { city: 'Melbourne', country: 'Australia' } }
+ * set('name', 'Steve', person) // { name: 'Steve', age: 20, address: { city: 'Melbourne', country: 'Australia' } }
+ * set('address.city', 'Sydney', person) // { name: 'Jane', age: 20, address: { city: 'Sydney', country: 'Australia' } }
  */
 export default curry(set)