+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Update `get` function to accept a key path
 * Remove `getIn` function
+* Update `set` function to accept a key path
 
 ## 3.0.2 (2019-04-09)